requirements.txt
langchain
ipykernel
langchain_community
python-dotenv
pypdf
langchain-text-splitters
sentence_transformers
streamlit